backend/requirements.txt
Django
MySQL-python
social-auth-app-django
django-environ
django-extensions
django-cities